deploy/requirements.txt
psycopg2-binary==2.9.10
langchain-core==0.3.66
langchain-anthropic==0.3.16
langchain-openai==0.3.27
langchain-postgres==0.0.15
langchain==0.3.26
unstructured[all-docs]==0.18.1
opencv-python-headless==4.12.0.88
awslambdaric